Refactor - Changed company address from string to separate entity.
